pyproject.toml
[tool.poetry]
name = "lms-core"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]

[tool.poetry.dependencies]
python = "^3.9"
Django = "^4.0.4"
django-environ = "^0.8.1"
ariadne = "0.13.0"
graphql-utils = "^0.4"
pylint-django = "^2.5.3"
pytest-django = "^4.5.2"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
